requirements.txt
django ==4.2.2
djangorestframework ==3.14.0
psycopg2-binary ==2.9.6
django-filter ==23.2
gunicorn
django-environ==0.11.2
requests


